Resolve 15 conflicts from merging master (PR #88 financeiro CAP/CAR +
payment orchestration + fiscal + response compression) into helpdesk
E2E branch. All changes from both sides kept:
- .csproj: all PackageReference/ProjectReference from both sides
- ServiceCollectionExtensions: all use case registrations merged
- DbContext: all DbSets from both sides (Payment, CAP/CAR, FAQ, Fiscal)
- ModelSnapshot: taken from master (superset of HEAD entities)
- IEfiPixService: kept PUT /v2/cob/{txid} (correct HTTP method)
- TicketDTOs + AbrirTicketUseCase: PedidoId support from master
- Permissao enum: all values from both sides
- Web/Program.cs: TicketsApiService + response compression from master

+++ EasyStock.Application/Ports/Output/Persistence/IMovimentacaoEstoqueRepository.cs
@@ -3,7 +3,7 @@ using EasyStock.Domain.Enums;
 
 namespace EasyStock.Application.Ports.Output.Persistence
 {
-    public sealed record KpisMovimentacao(int TotalUnidades, decimal ReceitaTotal, int TotalVendas, int TotalPerdas);
+    public sealed record KpisMovimentacao(decimal TotalUnidades, decimal ReceitaTotal, int TotalVendas, int TotalPerdas);
 
     public interface IMovimentacaoEstoqueRepository
     {
